+++ src/MicroElements.AutoFixture.NodaTime/LocalDateGenerator.cs
@@ -1,19 +1,29 @@
-﻿using System;
+﻿// Copyright (c) MicroElements. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+using System;
 using AutoFixture.Kernel;
 using NodaTime;
 
 namespace MicroElements.AutoFixture.NodaTime
 {
+    /// <summary>
+    /// AutoFixture generator for <see cref="LocalDate"/>.
+    /// </summary>
     public class LocalDateGenerator : ISpecimenBuilder
     {
         /// <inheritdoc />
         public object Create(object request, ISpecimenContext context)
         {
             if (context == null)
+            {
                 throw new ArgumentNullException(nameof(context));
+            }
 
             if (!typeof(LocalDate).Equals(request))
+            {
                 return new NoSpecimen();
+            }
 
             return LocalDate.FromDateTime(DateTime.Today);
         }