+++ src/MicroElements.AutoFixture.NodaTime/LocalDateTimeGenerator.cs
@@ -1,11 +1,18 @@
-﻿using System;
+﻿// Copyright (c) MicroElements. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+using System;
 using AutoFixture.Kernel;
 using NodaTime;
 
 namespace MicroElements.AutoFixture.NodaTime
 {
+    /// <summary>
+    /// AutoFixture generator for <see cref="LocalDateTime"/>.
+    /// </summary>
     public class LocalDateTimeGenerator : ISpecimenBuilder
     {
+        /// <inheritdoc />
         public object Create(object request, ISpecimenContext context)
         {
             if (context == null)