+++ src/MicroElements.AutoFixture.NodaTime/LocalTimeGenerator.cs
@@ -1,9 +1,15 @@
-﻿using System;
+﻿// Copyright (c) MicroElements. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+using System;
 using AutoFixture.Kernel;
 using NodaTime;
 
 namespace MicroElements.AutoFixture.NodaTime
 {
+    /// <summary>
+    /// AutoFixture generator for <see cref="LocalTime"/>.
+    /// </summary>
     public class LocalTimeGenerator : ISpecimenBuilder
     {
         /// <inheritdoc />