+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 0.2.0
+- Added InstantGenerator (PR# 1 by  sfmskywalker)
+- NodaTime updated to 2.4.7
+- Added reference to AutoFixture.Xunit2
+- Added NodaTimeAutoDataAttribute to use with xUnit
+
 # 0.1.0
 - Initial version